cgame: no need for explicit casting

+++ src/cgame/cg_newDraw.c
@@ -481,7 +481,7 @@ void CG_DrawCursorhint(rectDef_t *rect)
 
 	if (cg_cursorHints.integer == 3)
 	{
-		color[3] *= 0.5 + 0.5 * sin((double)cg.time / 150.0);
+		color[3] *= 0.5 + 0.5 * sin(cg.time / 150.0);
 	}
 
 	// size
@@ -493,11 +493,11 @@ void CG_DrawCursorhint(rectDef_t *rect)
 	{
 		if (cg_cursorHints.integer == 2)
 		{
-			scale = (float)((cg.cursorHintTime) % 1000) / 100.0f;     // one way size pulse
+			scale = (cg.cursorHintTime % 1000) / 100.0f;     // one way size pulse
 		}
 		else
 		{
-			scale = (float)(CURSORHINT_SCALE * (0.5 + 0.5 * sin((double)cg.time / 150.0)));     // sin pulse
+			scale = (float)(CURSORHINT_SCALE * (0.5 + 0.5 * sin(cg.time / 150.0)));     // sin pulse
 
 		}
 		halfscale = scale * 0.5f;
